Extract binding callback context
